+++ HeroFreePick/MenuLayout.lua
@@ -1,9 +1,9 @@
 HeroMenuLayout={frames={},defaults={},overrides=HeroMenuLayoutOverrides or {}}
 local L=HeroMenuLayout
 L.defaults['window']={parent='',x=0,y=0,w=1180,h=620}
-L.defaults['header']={parent='window',x=48,y=7,w=1124,h=29}
-L.defaults['close']={parent='window',x=1143,y=7,w=29,h=29}
-L.defaults['info']={parent='window',x=1114,y=7,w=29,h=29}
+L.defaults['header']={parent='window',x=48,y=7,w=1031,h=29}
+L.defaults['close']={parent='window',x=1147,y=11,w=22,h=22}
+L.defaults['info']={parent='window',x=1119,y=11,w=22,h=22}
 L.defaults['abilities']={parent='window',x=25,y=135,w=451,h=433}
 L.defaults['talents']={parent='window',x=483,y=135,w=334,h=433}
 L.defaults['side']={parent='window',x=828,y=106,w=326,h=462}